% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{approx_theta_post_s2}
\alias{approx_theta_post_s2}
\title{Find optimal theta of a CNV segment using forward-backward; uses a 2-state allele HMM}
\usage{
approx_theta_post_s2(
  pAD,
  DP,
  R,
  p_s,
  upper = 0.499,
  start = 0.25,
  gamma = 20,
  r = 0.015
)
}
\description{
Find optimal theta of a CNV segment using forward-backward; uses a 2-state allele HMM
}
\keyword{internal}
