% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{approx_theta_post_s3}
\alias{approx_theta_post_s3}
\title{Find optimal theta of a chromosome using forward-backward; uses a 3-state allele HMM}
\usage{
approx_theta_post_s3(
  pAD,
  DP,
  R,
  p_s,
  t,
  upper = 0.45,
  start = 0.05,
  gamma = 20,
  r = 0.015
)
}
\description{
Find optimal theta of a chromosome using forward-backward; uses a 3-state allele HMM
}
\keyword{internal}
