% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{calc_allele_lik_s2}
\alias{calc_allele_lik_s2}
\title{Calculate allele likelihoods for 2-state allele HMM}
\usage{
calc_allele_lik_s2(pAD, DP, R, p_s, theta, gamma = 20, r = 0.015)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{R}{numeric vector Variant mapping bias direction}

\item{p_s}{numeric vector Phase switch probabilities}

\item{theta}{numeric Haplotype imbalance}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\value{
numeric vector Allele likelihoods
}
\description{
Calculate allele likelihoods for 2-state allele HMM
}
\keyword{internal}
