% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{calc_allele_lik_s3}
\alias{calc_allele_lik_s3}
\title{Calculate allele likelihoods for 3-state allele HMM}
\usage{
calc_allele_lik_s3(pAD, DP, R, p_s, t, theta, gamma = 20, r = 0.015)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{p_s}{numeric vector Phase switch probabilities}

\item{theta}{numeric Haplotype imbalance}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\description{
Calculate allele likelihoods for 3-state allele HMM
}
\keyword{internal}
