% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{calc_trans_mat_s7}
\alias{calc_trans_mat_s7}
\title{Calculate the transition matrix for 7-state joint HMM}
\usage{
calc_trans_mat_s7(t, p_s, w, states_cn, states_phase)
}
\arguments{
\item{t}{numeric; CNV state transition probability}

\item{p_s}{numeric; phase switch probability}

\item{w}{numeric; relative abundance of states}

\item{states_cn}{character; CNV states}

\item{states_phase}{character; haplotype phase states}
}
\value{
array; transition matrix
}
\description{
Calculate the transition matrix for 7-state joint HMM
}
\keyword{internal}
