% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{check_cols}
\alias{check_cols}
\title{Check if columns are present in dataframe}
\usage{
check_cols(df, cols)
}
\arguments{
\item{df}{dataframe Dataframe}

\item{cols}{character vector Column names}
}
\value{
dataframe Dataframe
}
\description{
Check if columns are present in dataframe
}
\keyword{internal}
