% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_lnpois_cpp}
\alias{fit_lnpois_cpp}
\title{Fit MLE of log-normal Poisson model}
\usage{
fit_lnpois_cpp(Y_obs, lambda_ref, d)
}
\arguments{
\item{Y_obs}{Vector of observed counts}

\item{lambda_ref}{Vector of reference rates}

\item{d}{integer Total depth}
}
\value{
NumericVector MLE estimates of mu and sigma
}
\description{
Fit MLE of log-normal Poisson model
}
