% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts_example}
\alias{gene_counts_example}
\title{example gene expression counts matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1758 rows and 1 columns.
}
\usage{
gene_counts_example
}
\description{
example gene expression counts matrix
}
\keyword{datasets}
