% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_allele_bulk}
\alias{get_allele_bulk}
\title{Aggregate into pseudobulk alelle profile}
\usage{
get_allele_bulk(df_allele, gtf, genetic_map = NULL, nu = 0.5, min_depth = 0)
}
\arguments{
\item{df_allele}{dataframe Single-cell allele counts}

\item{gtf}{dataframe Transcript gtf}

\item{genetic_map}{dataframe Genetic map}

\item{nu}{numeric Phase switch rate}

\item{min_depth}{integer Minimum coverage to filter SNPs}
}
\value{
dataframe Pseudobulk allele profile
}
\description{
Aggregate into pseudobulk alelle profile
}
\examples{
bulk_example = get_allele_bulk(
    df_allele = df_allele_example,
    gtf = gtf_hg38)
}
