% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{get_allele_hmm_s2}
\alias{get_allele_hmm_s2}
\title{Make a 2-state allele HMM - no transitions to netural state}
\usage{
get_allele_hmm_s2(pAD, DP, R, p_s, theta, gamma = 20, r = 0.015)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{p_s}{numeric vector Phase switch probabilities}

\item{theta}{numeric Haplotype imbalance}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\value{
HMM object
}
\description{
Make a 2-state allele HMM - no transitions to netural state
}
\keyword{internal}
