% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{get_allele_hmm_s3}
\alias{get_allele_hmm_s3}
\title{Make a 3-state allele HMM - allow transitions to netural state}
\usage{
get_allele_hmm_s3(pAD, DP, R, p_s, t, theta_min, gamma = 20, r = 0.015)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{p_s}{numeric vector Phase switch probabilities}

\item{t}{numeric Transition probability between haplotype states}

\item{theta_min}{numeric Haplotype imbalance}

\item{gamma}{numeric Overdispersion in the allele-specific expression}

\item{r}{numeric Variant mapping bias}
}
\value{
HMM object
}
\description{
Make a 3-state allele HMM - allow transitions to netural state
}
\keyword{internal}
