% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{get_trans_probs_s15}
\alias{get_trans_probs_s15}
\title{Helper function to calculate transition porbabilities for 15-state joint HMM
cn/phase are sclars, only p_s is vectorized}
\usage{
get_trans_probs_s15(t, p_s, w, cn_from, phase_from, cn_to, phase_to)
}
\arguments{
\item{t}{numeric; CNV state transition probability}

\item{p_s}{numeric; phase switch probability}

\item{w}{numeric; relative abundance of states}

\item{cn_from}{character; originating CNV state}

\item{phase_from}{character; originating haplotype phase state}

\item{cn_to}{character; destination CNV state}

\item{phase_to}{character; destination haplotype phase state}
}
\value{
numeric; transition probability
}
\description{
Helper function to calculate transition porbabilities for 15-state joint HMM
cn/phase are sclars, only p_s is vectorized
}
\keyword{internal}
