% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{l_lnpois}
\alias{l_lnpois}
\title{calculate joint likelihood of a PLN model}
\usage{
l_lnpois(Y_obs, lambda_ref, d, mu, sig, phi = 1)
}
\arguments{
\item{Y_obs}{numeric vector Gene expression counts}

\item{lambda_ref}{numeric vector Reference expression levels}

\item{d}{numeric Total library size}

\item{mu}{numeric Global mean expression}

\item{sig}{numeric Global standard deviation of expression}

\item{phi}{numeric Fold change of expression}
}
\value{
numeric Joint log likelihood
}
\description{
calculate joint likelihood of a PLN model
}
\examples{
l_lnpois(c(1, 2), c(1, 2), 1, 1, 1)
}
