% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{run_allele_hmm_s3}
\alias{run_allele_hmm_s3}
\title{Run a 3-state allele HMM}
\usage{
run_allele_hmm_s3(
  pAD,
  DP,
  R,
  p_s,
  t = 1e-05,
  theta_min = 0.08,
  gamma = 20,
  r = 0.015
)
}
\arguments{
\item{pAD}{integer vector Paternal allele counts}

\item{DP}{integer vector Total alelle counts}

\item{R}{numeric vector Variant mapping bias direction}

\item{p_s}{numeric vector Phase switch probabilities}

\item{t}{numeric Transition probability between copy number states}

\item{theta_min}{numeric Minimum haplotype frequency deviation threshold}

\item{gamma}{numeric Overdispersion in the allele-specific expression}
}
\value{
character vector Decoded states
}
\description{
Run a 3-state allele HMM
}
\keyword{internal}
