% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{viterbi_joint}
\alias{viterbi_joint}
\title{Viterbi algorithm for joint HMM, can only handle one set of observations}
\usage{
viterbi_joint(hmm)
}
\arguments{
\item{hmm}{HMM object; expect variables x (allele depth), d (total depth), y (expression count), l (cell total library size),
lambda (reference expression rate), mu (global expression mean), sig (global expression standard deviation),
logPi (log transition prob matrix), phi (expression fold change for each state), delta (prior for each state),
alpha (alpha for each state), beta (beta for each state), states (states), p_s (phase switch probs)}
}
\value{
character vector Decoded states
}
\description{
Viterbi algorithm for joint HMM, can only handle one set of observations
}
\keyword{internal}
