% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hahmmr.R
\name{viterbi_joint_mat}
\alias{viterbi_joint_mat}
\title{Generalized viterbi algorithm for joint HMM, can handle multiple sets of observations}
\usage{
viterbi_joint_mat(hmm)
}
\arguments{
\item{hmm}{HMM object; expect variables x (allele depth), d (total depth), y (expression count), l (cell total library size),
lambda (reference expression rate), mu (global expression mean), sig (global expression standard deviation),
logPi (log transition prob matrix), phi (expression fold change for each state), delta (prior for each state),
alpha (alpha for each state), beta (beta for each state), states (states), p_s (phase switch probs)}
}
\description{
Generalized viterbi algorithm for joint HMM, can handle multiple sets of observations
}
\keyword{internal}
